% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gdp_v_lifeexp}
\alias{plot_gdp_v_lifeexp}
\title{Plot mean life expectancy against mean GDP}
\usage{
plot_gdp_v_lifeexp(df, palette = c("viridis", "magma", "inferno", "plasma"))
}
\arguments{
\item{df}{yearly means data.frame as output by \code{summarise_by_year}}

\item{palette}{specify the viridis palette option}
}
\description{
Plot mean life expectancy against mean GDP
}
